Partially fixed windows system specific functions. Done: - Directory control - Virtual File System support.
Missing:
- Specific Windows timing functions.
- Interrupt.
- Specific Windows file stream.

+++ nengine/engine/datawriter.cpp
@@ -238,7 +238,7 @@ bool DataWriter::write( double value )
 
 bool DataWriter::write( const String& tgt )
 {
-   byte nCharCount=tgt.manipulator()->charSize();
+   byte nCharCount=(byte) tgt.manipulator()->charSize();
    length_t size = tgt.size();
 
    write( nCharCount );
